Norminette - Splitting the functions so they are norminette valid (less than 25 lines) WIP - Leaving comments for now in case of debugging
