+++ lemin_files/ft_ants_to_paths.c
@@ -6,82 +6,11 @@
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:11:44 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/21 10:14:06 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/07/29 11:42:19 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/lemin.h"
-
-/*
-** Precalculation of the number of lines we are going to need to move
-** all the ants from source room to sink.
-** The final result lines is max_lines - 2 because in the path_size
-** source room and sink room are included but while moving the ants
-** we do not need to include them in the lines.
-** The reason is because in the beginning the subject assumes that
-** all ants are at the source room so we start printing lines from the
-** next room. Here we save 1 line.
-** Also, as soon as the last ant reached the end we are over so we do not
-** an extra line to move the ant from sink to somewhere else. And here we save
-** the second line.
-*/
-
-void				ft_calculate_lines(t_ant_farm *ant_farm)
-{
-	t_paths			*paths;
-	size_t			total_size;
-	size_t			max_lines;
-
-	max_lines = 0;
-	paths = ant_farm->paths;
-	// ft_printf(">> Calculate Lines - Current Run : %d \n", ant_farm->best_run);
-	while (paths != NULL) // && paths->run == ant_farm->best_run)
-	{
-		if (paths->ants_lst != NULL && paths->run == ant_farm->best_run)
-		{
-			// ft_printf("...path_id %d, paths_size %d, path run %d\n", paths->path_id, paths->path_size, paths->run);
-			// ft_printf("path_id %d, paths_size %d, ants %d\n", paths->path_id, paths->path_size, paths->ants_amount);
-			total_size = paths->path_size + paths->ants_amount;
-			if (total_size > max_lines)
-				max_lines = total_size;
-		}
-		paths = paths->next;
-	}
-	ant_farm->lines = max_lines - 2;
-	// ft_printf(ANSI_COLOR_CYAN"Lines Needed: %d \n\n", ant_farm->lines);
-}
-
-void				ft_ants_to_pathlst_addend(t_ants **lst, t_ants *new)
-{
-	t_ants			*temp;
-
-	if (*lst == NULL)
-	{
-		*lst = new;
-		return ;
-	}
-	temp = *lst;
-	while (temp->next != NULL)
-		temp = temp->next;
-	temp->next = new;
-	new->prev = temp;
-}
-
-void				ft_lines_list_addend(t_lines **lst, t_lines *new)
-{
-	t_lines			*temp;
-
-	if (*lst == NULL)
-	{
-		*lst = new;
-		return ;
-	}
-	temp = *lst;
-	while (temp->next != NULL)
-		temp = temp->next;
-	temp->next = new;
-	new->prev = temp;
-}
 
 int					ft_all_ants_in_same_path(t_ant_farm *ant_farm)
 {
@@ -100,17 +29,48 @@ int					ft_all_ants_in_same_path(t_ant_farm *ant_farm)
 		ant->ant_id = i;
 		ant->path = tmp;
 		// ft_printf(ANSI_COLOR_BLUE"ant_id: %d in path_id: %d\n"ANSI_COLOR_RESET, ant->ant_id, ant->path->path_id);
-		ft_ants_to_pathlst_addend(&(tmp->ants_lst), ant);
+		ft_add_ant_to_antslist(&(tmp->ants_lst), ant);
 		i += 1;
 	}
 	tmp->ants_amount = i - 1;
 	return (SUCCESS);
 }
 
+void				ft_assign_based_on_size(t_ants **ant, t_paths **cur, \
+											t_paths *first, t_paths *next)
+{
+	size_t			first_size;
+	size_t			next_size;
+
+	first_size = first->path_size + first->ants_amount;
+	next_size = next->path_size + next->ants_amount;
+	if (((*cur)->path_size + (*cur)->ants_amount <= next_size) && \
+	((*cur)->path_size + (*cur)->ants_amount <= first_size))
+	{
+		// assign to current path
+		(*ant)->path = (*cur);
+		(*cur)->ants_amount += 1;
+	}
+	else if (((*cur)->path_size + (*cur)->ants_amount <= next_size) && \
+	((*cur)->path_size + (*cur)->ants_amount > first_size))
+	{
+		// assign to first path
+		(*ant)->path = first;
+		first->ants_amount += 1;
+		(*cur) = first;
+	}
+	else
+	{
+		// assign ant to next path
+		(*ant)->path = next;
+		next->ants_amount += 1;
+		(*cur) = next;
+	}
+}
+
 /*
-** Comparing the total_size (path_size + ants_amount) of current_path,
-** next_path and first_path and we decide to assign the ant to the
-** current, next or first path respectively.
+** We compare the total_size (path_size + ants_amount) of current_path,
+** next_path and first_path and we decide in which path to assign the ant.
 */
 
 void				ft_compare_sizes(t_ant_farm *ant_farm, t_paths **cur, \
@@ -119,8 +79,6 @@ void				ft_compare_sizes(t_ant_farm *ant_farm, t_paths **cur, \
 	t_paths			*first;
 	t_paths			*next;
 	t_ants			*ant;
-	size_t			first_size;
-	size_t			next_size;
 
 	ant = (t_ants *)ft_memalloc(sizeof(t_ants));
 	ant->ant_id = i;
@@ -128,7 +86,6 @@ void				ft_compare_sizes(t_ant_farm *ant_farm, t_paths **cur, \
 	while (first != NULL && first->run != ant_farm->best_run)
 		first = first->next;
 	// ft_printf(">> Compare sizes best run %d next \n", ant_farm->best_run);
-	first_size = first->path_size + first->ants_amount;
 	if ((*cur)->next != NULL && (*cur)->next->run == ant_farm->best_run)
 		next = (*cur)->next;
 	else
@@ -137,20 +94,6 @@ void				ft_compare_sizes(t_ant_farm *ant_farm, t_paths **cur, \
 		while (next != NULL && next->run != ant_farm->best_run)
 			next = next->next;
 	}
-	// if ((*cur)->next != NULL && (*cur)->next->run == ant_farm->best_run)
-	// 	next = (*cur)->next;
-	// else if ((*cur)->next != NULL && (*cur)->next->run != ant_farm->best_run)
-	// {
-	// 	while ((*cur)->next != NULL && (*cur)->next->run != ant_farm->best_run)
-	// 		next = (*cur)->next;
-	// }
-	// else if ((*cur)->next == NULL)
-	// {
-	// 	next = ant_farm->paths;
-	// 	while (next != NULL && next->run != ant_farm->best_run)
-	// 		next = next->next;
-	// }
-	next_size = next->path_size + next->ants_amount;
 	// ft_printf(ANSI_COLOR_BLUE"cur_pathid: %d, size = %d\n"ANSI_COLOR_RESET, 
 	// (*cur)->path_id, (*cur)->path_size + (*cur)->ants_amount);
 	// if ((*cur)->next != NULL && (*cur)->next->run == ant_farm->best_run)
@@ -163,49 +106,25 @@ void				ft_compare_sizes(t_ant_farm *ant_farm, t_paths **cur, \
 			// ft_printf(ANSI_COLOR_YELLOW_PAST"next_pathid: %d, size = %d\n" 
 			// ANSI_COLOR_RESET, next->path_id, next_size);
 	// }
-	if (((*cur)->path_size + (*cur)->ants_amount <= next_size) && \
-	((*cur)->path_size + (*cur)->ants_amount <= first_size))
-	{
-		// assign to current path
-		ant->path = (*cur);
-		(*cur)->ants_amount += 1;
-	}
-	else if (((*cur)->path_size + (*cur)->ants_amount <= next_size) && \
-	((*cur)->path_size + (*cur)->ants_amount > first_size))
-	{
-		// assign to first path
-		ant->path = first;
-		first->ants_amount += 1;
-		(*cur) = first;
-	}
-	else
-	{
-		// assign ant to next path
-		ant->path = next;
-		next->ants_amount += 1;
-		(*cur) = next;
-	}
-	ft_ants_to_pathlst_addend(&(ant->path->ants_lst), ant);
-	// ft_printf(ANSI_COLOR_GREEN_EMER"ant_id: %d, path_id = %d\n\n"ANSI_COLOR_RESET, ant->ant_id, ant->path->path_id);
-	// return (ant);
+	ft_assign_based_on_size(&ant, cur, first, next);
+	ft_add_ant_to_antslist(&(ant->path->ants_lst), ant);
+	// ft_printf("ant_id:%d,path_id=%d\n\n", ant->ant_id, ant->path->path_id);
 }
 
 int					ft_ants_to_paths(t_ant_farm *ant_farm)
 {
 	t_paths			*tmp;
-	// t_path_list		*tmp_lst;
 	size_t			i;
 
 	// ft_printf("\nAssigning Ants to Paths\n");
-	// ft_printf("========================\n");
 	// ft_printf("Number of ants : %d \n", ant_farm->ants);
 	// ft_printf("Max paths : %d \n", ant_farm->max_paths);
 	// ft_printf("Discovered paths : %d \n", ant_farm->discovered_paths);
 	// ft_print_paths_list(ant_farm);
-	if (ant_farm->ants == 0)
+	if ((ant_farm->ants == 0) || (ant_farm->discovered_paths == 0))
 		return (SUCCESS);
-	if (ant_farm->discovered_paths == 0)
-		return (SUCCESS);
+	// if (ant_farm->discovered_paths == 0)
+	// 	return (SUCCESS);
 	if ((ant_farm->discovered_paths == 1) || (ant_farm->ants == 1))
 	{
 		// ft_printf("all ants in same path \n");
