+++ lemin_files/ft_bfs_sink.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   ft_bfs.c                                           :+:    :+:            */
+/*   ft_bfs_sink.c                                      :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:12:05 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/22 20:04:21 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/07/29 13:26:23 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
