+++ lemin_files/ft_input_parsing.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   ft_save_input.c                                    :+:    :+:            */
+/*   ft_input_parsing.c                                 :+:    :+:            */
 /*                                                     +:+                    */
-/*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
+/*   By: dsaripap <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
-/*   Created: 2020/06/19 10:39:21 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/06 17:31:20 by dsaripap      ########   odam.nl         */
+/*   Created: 2020/07/29 15:04:20 by dsaripap      #+#    #+#                 */
+/*   Updated: 2020/07/29 15:04:22 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,7 +108,7 @@ t_prgm_signal		ft_saveinput(t_ant_farm *ant_farm, char *line, size_t *j)
 		// ft_printf(ANSI_COLOR_CYAN"Invalid command\n"ANSI_COLOR_RESET);
 		return (CONTINUE);
 	}
-	else if (lm_check_if_ants_amount(ant_farm, line, *j) != CONTINUE)
+	else if (ft_check_if_ants_amount(ant_farm, line, *j) != CONTINUE)
 		return (ant_farm->signal);
 	else if (ft_check_if_is_room(ant_farm, line, link) != CONTINUE)
 		return (ant_farm->signal);
