+++ lemin_files/ft_map_data_functions.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   ft_save_inputlines.c                               :+:    :+:            */
+/*   ft_map_data_functions.c                            :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:39:29 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/06 17:30:27 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/07/29 20:25:33 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,4 +44,45 @@ void				ft_input_addend(t_input **lst, t_input *new)
 	while (temp->next != NULL)
 		temp = temp->next;
 	temp->next = new;
+}
+
+/*
+** Function that prints the linkedlist with the raw input lines
+** that we retrieved from the map
+*/
+
+void				print_input_list(t_input *input_data)
+{
+	t_input			*temp;
+
+	temp = input_data;
+	ft_printf("\n   Linked List with Raw Input Lines      \n");
+	ft_printf("-----------------------------------------\n");
+	ft_printf("|line\t\t\tnext_line\t|\n");
+	while (temp != NULL)
+	{
+		ft_printf("|%s\t\t", temp->line);
+		if (temp->next != NULL)
+			ft_printf("\t-> %s\t|\n", temp->next->line);
+		else
+			ft_printf("\t-> NULL\t|\n");
+		temp = temp->next;
+	}
+	ft_printf("----------------------------------------\n");
+}
+
+void			ft_print_mapdata(t_ant_farm *ant_farm)
+{
+	t_input		*lines;
+
+	lines = ant_farm->input;
+	while (lines != NULL)
+	{
+		// if ((lines->line[0] == '#') && (lines->line[1] != '#'))
+		// 	;
+		// else
+		ft_printf("%s\n", lines->line);
+		lines = lines->next;
+	}
+	ft_printf("\n");
 }
