+++ lemin_files/ft_min_cut_sink.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   ft_min_cut.c                                       :+:    :+:            */
+/*   ft_min_cut_sink.c                                  :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: dsaripap <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
-/*   Created: 2020/07/13 08:23:30 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/22 19:32:24 by dsaripap      ########   odam.nl         */
+/*   Created: 2020/07/29 15:21:04 by dsaripap      #+#    #+#                 */
+/*   Updated: 2020/07/29 15:21:09 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
